PR changes + app-header reorganized
